+++ test/unit/fuzz-tools.test.js
@@ -345,6 +345,7 @@ import * as pragueBarrettV365 from '../../lib/prague-barrett-v365.js';
 import * as neckZoneV366 from '../../lib/neck-zone-v366.js';
 import * as pasSwallowV367 from '../../lib/pas-swallow-v367.js';
 import * as rossHfPedsV368 from '../../lib/ross-hf-peds-v368.js';
+import * as nohriaStevensonV369 from '../../lib/nohria-stevenson-v369.js';
 import * as tbTesting from '../../lib/tb-testing.js';
 
 const MODULES = {
@@ -635,6 +636,7 @@ const MODULES = {
   'neck-zone-v366.js': neckZoneV366,
   'pas-swallow-v367.js': pasSwallowV367,
   'ross-hf-peds-v368.js': rossHfPedsV368,
+  'nohria-stevenson-v369.js': nohriaStevensonV369,
   'tb-testing.js': tbTesting,
 };